% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{abbreviate_datasets}
\alias{abbreviate_datasets}
\title{Abbreviate fly connectomics dataset names}
\usage{
abbreviate_datasets(ds)
}
\arguments{
\item{ds}{One or more dataset long names}
}
\value{
a character vector of 2 letter abbreviations
}
\description{
Abbreviate fly connectomics dataset names
}
\examples{
abbreviate_datasets(c("flywire", "flywire", "hemibrain"))
}
