% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{cf_datasets}
\alias{cf_datasets}
\title{List connectomics datasets known to coconatfly}
\usage{
cf_datasets(rval = c("all", "available", "builtin", "external"))
}
\arguments{
\item{rval}{Character string specifying which datasets to return. One of:
\itemize{
  \item \code{"all"} (default) all known datasets (builtin + externally
    registered)
  \item \code{"builtin"} only the datasets hardcoded into the package
  \item \code{"external"} only datasets registered via the
    \code{\link[coconat]{register_dataset}} mechanism
}}
}
\value{
A character vector of dataset names.
}
\description{
List connectomics datasets known to coconatfly
}
\examples{
cf_datasets()
cf_datasets('builtin')
}
\seealso{
\code{\link{abbreviate_datasets}}, \code{\link{cf_ids}}
}
