% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{cf_ids}
\alias{cf_ids}
\title{Specify ids for flybrain datasets}
\usage{
cf_ids(
  query = NULL,
  datasets = c("brain", "vnc", "hemibrain", "flywire", "malecns", "manc", "fanc"),
  hemibrain = NULL,
  flywire = NULL,
  malecns = NULL,
  manc = NULL,
  fanc = NULL
)
}
\arguments{
\item{query}{A query (e.g. cell type name or regular expression)}

\item{datasets}{Character vector naming datasets to which the \code{query}
should be applied.}

\item{hemibrain}{Pass hemibrain specific query or ids to this argument}

\item{flywire}{Pass flywire specific query or ids to this argument}

\item{malecns}{Pass malecns specific query or ids to this argument}

\item{manc}{Pass manc specific query or ids to this argument}

\item{fanc}{Pass fanc ids to this argument (at present we do not support
metadata queries for fanc)}
}
\value{
A list of ids
}
\description{
Specify ids for flybrain datasets
}
\examples{
\donttest{
cf_ids("DA2_lPN", datasets='brain')
# / introduces a regular expression
cf_ids("/MBON.+", datasets='brain')
}
}
\seealso{
Other ids: 
\code{\link{keys2df}()}
}
\concept{ids}
