% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners.R
\name{cf_partners}
\alias{cf_partners}
\title{Flexible function for fetching partner data across datasets}
\usage{
cf_partners(
  ids,
  threshold = 1L,
  partners = c("inputs", "outputs"),
  bind.rows = TRUE
)
}
\arguments{
\item{ids}{A list of ids named by the relevant datasets}

\item{threshold}{return only edges with at least this many matches}

\item{partners}{Whether to return inputs or outputs}

\item{bind.rows}{Whether to bind data.frames for each dataset together,
keeping only the common columns (default \code{TRUE} for convenience but
note that some columns will be dropped).}
}
\description{
Flexible function for fetching partner data across datasets
}
\examples{
\dontrun{
cf_partners(list(flywire='DA2_lPN'))

DA2_lPN=cf_partners(list(flywire='DA2_lPN', malecns='DA2_lPN'))
}
}
