% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfb.R
\name{cf_vfb_ids}
\alias{cf_vfb_ids}
\title{Find VFB ids for individual neurons}
\usage{
cf_vfb_ids(keys)
}
\arguments{
\item{keys}{A character vector of keys (like \code{hb:10001}) or a data frame
containing \code{id} and \code{dataset} columns.}
}
\value{
A data frame containing a \code{vfb_id} column if \code{keys} was a
  data frame, a character vector otherwise.
}
\description{
Find VFB ids for individual neurons
}
\examples{
\donttest{
dnm=cf_meta(cf_ids(manc='/type:DNa0[1-3]', hemibrain = '/type:DNa0[1-3]'))
cf_vfb_ids(mbonmeta)
# nb this function needs keys as input
cf_vfb_ids(cf_ids(manc='/type:DNa0[1-3]', keys=TRUE))
}
}
