% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconatfly-package.R
\docType{package}
\name{coconatfly-package}
\alias{coconatfly}
\alias{coconatfly-package}
\title{coconatfly: Comparative Connectomics of Public and In Progress Drosophila Datasets}
\description{
Experimental package to enable comparative connectomics analysis of multiple Drosophila connectomics datasets, including some private ones still under development.
}
\section{Package Options}{

\describe{
  \item{\code{coconatfly.use_superclass}}{Logical. If \code{TRUE}, rename
  class/subclass/subsubclass columns to superclass/class/subclass in
  \code{\link{cf_meta}}, \code{\link{cf_partners}}, and
  \code{\link{cf_partner_summary}}. Default: \code{FALSE}.
  Set with \code{options(coconatfly.use_superclass = TRUE)}.}
  \item{\code{coconatfly.harmonise_class}}{Logical. If \code{TRUE}, harmonise
  class values to malecns style across all datasets. Default: \code{FALSE}.
  Set with \code{options(coconatfly.harmonise_class = TRUE)}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/coconatfly}
  \item \url{https://natverse.org/coconatfly/}
  \item Report bugs at \url{https://github.com/natverse/coconatfly/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
