% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dr_coconatfly}
\alias{dr_coconatfly}
\title{Status report for coconatfly installation}
\usage{
dr_coconatfly()
}
\value{
A dataframe containing status information for connectome datasets (invisibly).
}
\description{
Status report for coconatfly installation
}
\examples{
\donttest{
dr_coconatfly()
}
}
