% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{keys}
\alias{keys}
\alias{keys2df}
\alias{keys2list}
\title{Interconvert between keys and ids/datasets}
\usage{
keys(x, idcol = "id")

keys2df(keys, integer64 = FALSE)

keys2list(keys, integer64 = FALSE)
}
\arguments{
\item{x}{A list, data frame, dendrogram, or character vector specifying both
within dataset ids and dataset names. See details and examples especially
for character vector input.}

\item{idcol}{optional string naming the column containing ids}

\item{keys}{A character vector of keys}

\item{integer64}{Whether the output ids should be character vectors (the
default) or \code{integer64}}
}
\value{
For \code{keys} a character vector of keys of the form
  \code{"<dataset>:<id>"}.
}
\description{
Neurons within a dataset will be identified by numeric ids but
  these may not be unique across datasets. Therefore to make identifiers that
  are unique across dataset we use \code{keys} of the form
  \code{"<dataset>:<id>"}.

  \code{keys} either confirms/tidies up an existing set of keys or converts a
  \code{list} or \code{data.frame} to keys.

\code{keys2df} produces a \code{data.frame} with columns
  \code{id} and \code{dataset} describing the ids for each dataset. The
  ordering of the data.frame will match the order of keys in the input
  vector.

\code{keys2list} converts a character vector of keys to a list of ids with one list element for each dataset
}
\details{
When \code{x} is a character vector, this must be in one of two
  forms. \emph{Either} a vector where each element is a single key of the
  form \code{"<dataset>:<id>"} \emph{or} a single string containing >=1 such
  keys separated by white space or commas (e.g. \code{"
  fw:4611686018427387904, hb:12345 "}). See examples.

  As a convenience \code{x} may also be a \code{dendrogram} or \code{hclust}
  object resulting from a clustering operation.
}
\examples{
# tidying up keys copied from somewhere else ...
keys(" fw:4611686018427387904, hb:12345 ")

\donttest{
keys(cf_ids(hemibrain=12345, flywire='4611686018427387904'))

# NB this runs the query for hemibrain type MBON01 and then maps ids -> keys
keys(cf_ids(hemibrain='MBON01'))
}
\donttest{
keys2df(cf_ids('MBON01', datasets = c("hemibrain", "flywire")))
}
\donttest{
keys2list(cf_ids('MBON01', datasets = c("hemibrain", "flywire")))
}
}
\seealso{
Other ids: 
\code{\link{cf_ids}()}
}
\concept{ids}
