% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{cf_cosine_plot}
\alias{cf_cosine_plot}
\title{Multidataset cosine clustering}
\usage{
cf_cosine_plot(
  ids,
  ...,
  threshold = 5,
  partners = c("outputs", "inputs"),
  labRow = "{type}_{abbreviate_datasets(dataset)}{side}",
  group = "type",
  heatmap = TRUE,
  interactive = FALSE,
  nas = c("zero", "drop"),
  method = c("ward.D", "single", "complete", "average", "mcquitty", "median", "centroid",
    "ward.D2")
)
}
\arguments{
\item{ids}{A list of ids named by the relevant datasets (see examples)}

\item{...}{Additional arguments passed to \code{\link{heatmap}}}

\item{threshold}{return only edges with at least this many matches. 0 is an
option since neuprint sometimes returns 0 weight edges.}

\item{partners}{Whether to return inputs or outputs}

\item{labRow}{A vector defining the row labels. If it is a character vector
length 1 containing any curly braces it will be interpreted as a
\code{\link{glue}} string for interpolation, using the data.frame returned
by \code{\link[neuprintr]{neuprint_get_meta}} as a source of information.}

\item{group}{Whether to group by cell \code{type} or another named column.}

\item{heatmap}{A logical indicating whether or not to plot the heatmap
\emph{OR} a function to plot the heatmap whose argument names are
compatible with \code{stats::\link{heatmap}}. \code{gplots::heatmap.2} is a
good example. Defaults to \code{TRUE} therefore plotting the full heatmap
with \code{stats::\link{heatmap}}.}

\item{interactive}{Whether to plot an interactive heatmap (allowing zooming
and id selection). See details.}

\item{nas}{What to do with entries that have NAs. Default is to set them to 0
similarity.}

\item{method}{The cluster method to use (see \code{\link{hclust}})}
}
\value{
The result of \code{\link{heatmap}} invisibly including the row and
  column dendrograms or when \code{heatmap=FALSE}, an \code{\link{hclust}}
  dendrogram
}
\description{
Multidataset cosine clustering
}
\examples{
cf_cosine_plot(list(flywire="/type:LAL.+", malecns="/type:LAL.+"))
}
