% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{cf_ids}
\alias{cf_ids}
\title{Specify ids for fly connectome datasets}
\usage{
cf_ids(
  query = NULL,
  datasets = c("brain", "vnc", "hemibrain", "flywire", "malecns", "manc", "fanc"),
  expand = FALSE,
  keys = FALSE,
  hemibrain = NULL,
  flywire = NULL,
  malecns = NULL,
  manc = NULL,
  fanc = NULL
)
}
\arguments{
\item{query}{A query (e.g. cell type name or regular expression)}

\item{datasets}{Character vector naming datasets to which the \code{query}
should be applied.}

\item{expand}{Whether to expand any queries into the matching ids (this will
involve one or more calls to corresponding servers). Default \code{FALSE}.}

\item{keys}{Whether to turn the ids into keys \code{hb:12345} right away.
Default \code{FALSE} but you may find this useful e.g. for combining
lists of neurons (see examples).}

\item{hemibrain}{Pass hemibrain specific query or ids to this argument}

\item{flywire}{Pass flywire specific query or ids to this argument}

\item{malecns}{Pass malecns specific query or ids to this argument}

\item{manc}{Pass manc specific query or ids to this argument}

\item{fanc}{Pass fanc ids to this argument (at present we do not support
metadata queries for fanc)}
}
\value{
A list of ids
}
\description{
Specify ids for fly connectome datasets
}
\examples{
\donttest{
cf_ids("DA2_lPN", datasets='brain')
# / introduces a regular expression
cf_ids("/MBON.+", datasets='brain')

# expand query into actual ids
cf_ids("/type:MBON.+", datasets='brain', expand=TRUE)

# return keys directly
cf_ids("/type:MBON.+", keys=TRUE)
# can be useful for combining separate lists of neurons
hbids=c(264083994, 5813022274)
c(cf_ids("/type:MBON1.+", keys=TRUE), cf_ids(hemibrain = hbids, keys = T))
}
}
\seealso{
Other ids: 
\code{\link{keys}()}
}
\concept{ids}
