% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{cf_meta}
\alias{cf_meta}
\title{Fetch metadata for neurons from connectome datasets}
\usage{
cf_meta(
  ids,
  bind.rows = TRUE,
  integer64 = FALSE,
  flywire_type = c("cell_type", "hemibrain_type")
)
}
\arguments{
\item{ids}{A list of ids named by the relevant datasets (see examples)}

\item{bind.rows}{Whether to bind data.frames for each dataset together,
keeping only the common columns (default \code{TRUE} for convenience but
note that some columns will be dropped).}

\item{integer64}{Whether ids should be character vectors (default) or 64 bit
ints (more compact but a little fragile as they rely on the \code{bit64}
extension package.)}

\item{flywire_type}{Which metadata column to use from flytable info table}
}
\description{
Fetch metadata for neurons from connectome datasets
}
\examples{
\donttest{
da2meta=cf_meta(list(hemibrain='DA2_lPN'))
da2meta
# / introduces a regular expression
mbonmeta=cf_meta(list(hemibrain='/MBON.+'))
}
}
\seealso{
\code{\link{neuprint_ids}}
}
