% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{cf_meta}
\alias{cf_meta}
\title{Fetch metadata for neurons from connectome datasets}
\usage{
cf_meta(
  ids,
  bind.rows = TRUE,
  integer64 = FALSE,
  keep.all = FALSE,
  MoreArgs = list(flywire = list(type = c("cell_type", "hemibrain_type")))
)
}
\arguments{
\item{ids}{A list of ids named by the relevant datasets (see examples) or any
other input that can be processed by the \code{\link{keys}} function
(including a \code{hclust} dendrogram object.)}

\item{bind.rows}{Whether to bind data.frames for each dataset together,
keeping only the common columns (default \code{TRUE} for convenience but
note that some columns will be dropped by unless \code{keep.all=TRUE}).}

\item{integer64}{Whether ids should be character vectors (default) or 64 bit
ints (more compact but a little fragile as they rely on the \code{bit64}
extension package.)}

\item{keep.all}{Whether to keep all columns when processing multiple datasets
rather than just those in common (default=\code{FALSE} only keeps shared
columns).}

\item{MoreArgs}{A named list of arguments to be passed when fetching metadata
for a given function. See details.}
}
\description{
Fetch metadata for neurons from connectome datasets
}
\details{
\code{MoreArgs} should be list named by the standard dataset names
  (e.g. as returned by \code{cf_datasets}.
}
\examples{
\donttest{
da2meta=cf_meta(cf_ids(hemibrain='DA2_lPN'))
da2meta
# / introduces a regular expression
mbonmeta=cf_meta(cf_ids(hemibrain='/MBON.+'))
}
}
\seealso{
\code{\link[neuprintr]{neuprint_ids}}
}
