% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partners.R
\name{cf_partners}
\alias{cf_partners}
\title{Flexible function for fetching partner data across datasets}
\usage{
cf_partners(
  ids,
  threshold = 1L,
  partners = c("inputs", "outputs"),
  bind.rows = TRUE,
  MoreArgs = list()
)
}
\arguments{
\item{ids}{A list of ids named by the relevant datasets (see examples) or any
other input that can be processed by the \code{\link{keys}} function
(including a \code{hclust} dendrogram object.)}

\item{threshold}{return only edges with at least this many matches. 0 is an
option since neuprint sometimes returns 0 weight edges.}

\item{partners}{Whether to return inputs or outputs}

\item{bind.rows}{Whether to bind data.frames for each dataset together,
keeping only the common columns (default \code{TRUE} for convenience but
note that some columns will be dropped).}

\item{MoreArgs}{Additional arguments in the form of a hierarchical list
(expert use; see details and examples).}
}
\value{
A data.frame or a named list (when \code{bind.rows=FALSE})
}
\description{
Flexible function for fetching partner data across datasets
}
\details{
fancr and fafbseg functions have usually used a \code{>}
  relationship for the threshold, but here (as of May 2024) it is uniformly a
  \code{>=} relationship.

  \code{MoreArgs} is structured as a list with a top layer naming datasets
  (using the same long names as \code{\link{cf_ids}}. The second (lower)
  layer names the arguments that will be passed to dataset-specific functions
  such as \code{fafbseg::flywire_partner_summary2} and
  \code{malevnc::manc_connection_table}.
}
\examples{
\donttest{
da2ds=cf_partners(list(hemibrain='DA2_lPN'), threshold=5)
library(dplyr)
da2ds \%>\%
  count(type, wt=weight, sort=TRUE)
}
\dontrun{
cf_partners(list(flywire='DA2_lPN'))

DA2_lPN=cf_partners(list(flywire='DA2_lPN', malecns='DA2_lPN'))

cf_partners(cf_ids(malecns='AVLP539'), partners = 'o', threshold=100)
# use MoreArgs to prefer foreign (flywire/manc) cell types for malecns
cf_partners(cf_ids(malecns='AVLP539'), partners = 'o', threshold=100,
  MoreArgs = list(malecns=list(prefer.foreign=TRUE))
}
}
