% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coconatfly-package.R
\docType{package}
\name{coconatfly-package}
\alias{coconatfly}
\alias{coconatfly-package}
\title{coconatfly: Comparative Connectomics of Public and In Progress Drosophila Datasets}
\description{
Experimental package to enable comparative connectomics analysis of multiple Drosophila connectomics datasets, including some private ones still under development.
}
\section{Package options}{


coconatfly itself currently only has one option, which is to use a historical
behaviour for collating cell types for multi-connectome cosine clustering.

\itemize{

\item \code{options(coconatfly.cluster.dropna)}

}
}

\examples{
# list any package options that have been set
options()[grepl("^coconatfly.*", names(options()))]

# return to behaviour of coconatfly < v0.2
options(coconatfly.cluster.dropna=FALSE)
# return to default behaviour of coconatfly >= v0.2
options(coconatfly.cluster.dropna=TRUE)

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/natverse/coconatfly}
  \item \url{https://natverse.org/coconatfly/}
  \item Report bugs at \url{https://github.com/natverse/coconatfly/issues}
}

}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

}
\keyword{internal}
