% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{keys2df}
\alias{keys2df}
\alias{keys2list}
\alias{keys}
\title{Interconvert between keys and ids/datasets}
\usage{
keys2df(keys, integer64 = FALSE)

keys2list(keys, integer64 = FALSE)

keys(x)
}
\arguments{
\item{keys}{A character vector of keys}

\item{integer64}{Whether the output ids should be character vectors (the
default) or \code{integer64}}

\item{x}{A list or dataframe specifying both within dataset ids and dataset
names.}
}
\value{
For \code{keys} as character vector of keys of the form
  \code{"<dataset>:<id>"}.
}
\description{
\code{keys2list} converts a character vector of keys to a list of ids with one list element for each dataset

Neurons within a dataset will be identified by numeric ids but
these may not be unique across datasets. Therefore to make a unique datatset
we use \code{keys} of the form \code{"<dataset>:<id>"}.
}
\examples{
\donttest{

}
\donttest{
keys(list(hemibrain=12345, flywire='4611686018427387904'))
}
}
