% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ids.R
\name{keys}
\alias{keys}
\alias{keys2df}
\alias{keys2list}
\title{Interconvert between keys and ids/datasets}
\usage{
keys(x)

keys2df(keys, integer64 = FALSE)

keys2list(keys, integer64 = FALSE)
}
\arguments{
\item{x}{A list or dataframe specifying both within dataset ids and dataset
names.}

\item{keys}{A character vector of keys}

\item{integer64}{Whether the output ids should be character vectors (the
default) or \code{integer64}}
}
\value{
For \code{keys} as character vector of keys of the form
  \code{"<dataset>:<id>"}.
}
\description{
Neurons within a dataset will be identified by numeric ids but
these may not be unique across datasets. Therefore to make a unique datatset
we use \code{keys} of the form \code{"<dataset>:<id>"}.

\code{keys2df} produces a \code{data.frame} with columns
  \code{id} and \code{dataset} describing the ids for each dataset. The
  ordering of the data.frame will match the order of keys in the input
  vector.

\code{keys2list} converts a character vector of keys to a list of ids with one list element for each dataset
}
\examples{
\donttest{
keys(cf_ids(hemibrain=12345, flywire='4611686018427387904'))
}
\donttest{

}
}
\seealso{
Other ids: 
\code{\link{cf_ids}()}
}
\concept{ids}
